+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add a non-SO wild items to the default config for feature and qualifier names
 * Fix: Respect feature-wise /transl_table value than the globally-set one.
 * Fix an error when joined features have .sub_features
+* Fix [source] in metadata not used when "source" feature exists in a entry
 
 ## v0.2.4 (2021-10-04)
 * Expand the default config for feature and qualifier names